requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
pillow
numpy
psycopg2-binary
pandas
matplotlib
